Introducing confidence intervals, according to the Student's T distribution.
Confidence intervals (for p=.90,.95,.99) are now supported and
displayed by the program. They are also printed to .csv files
when -f is specified.
A new option has been introduced (-C) to select which confidence
intervals to display (all of them are instead always written to
.csv files).
Standard deviation, as an additional statistic, is now computed too,
using Welford's algorithm.