Introducing hardware timestamping support (1st prototype)
Introducing the new 'h' latency type (-L h), which enables, on
supported platforms and with supported NICs, hardware timestamping.
This features follows the LaMP rev2.0 specifications and it is
currently using a mandatory follow-up mechanism.
It should be still considered as a prototype, as follow-up messages
are going to be exteneded to all the other latency types and they
will not be made as mandatory for the hardware timestamping latency
type.
Some debug messages are still being printed (they are marked by [TBR],
i.e. "To Be Removed") but they will be removed as the development
progresses.
Some code cleanup operations must still be performed.