-W now writes sequence numbers to CSV, removed rx_bytes per-packet statistic
-W now writes sequence numbers to the specified CSV file too, as it
makes it easier to track when packets are lost, if there are some
lost packets.
The per-packet statistic 'rx_bytes', which was printed for every received
reply in ping-like mode, has been removed. This has been done as it was
not very meaningful and it was also generating some wrong output when used
in conjunction with follow-up packets.