New development branch for v0.1.5-beta: set new version number
Setting up the new version number in version.h and README.md, as the development
branch is now for v0.1.5-beta-development and no more for v0.1.4-beta-development,
which has been just merged into master (releasing v0.1.4-beta).
This commit also restores the "development" tag after the version number, which was
removed for merging the development branch into the master one.