Fix Makefile to create obj folder and fix confidence intervals
This commit fixed the Makefile, as it was not creating the 'obj'
directory, resulting in an error when calling 'make'.
The most important modification, is, however, related to the
confidence intervals computation, which was referring to one-tailed
tests, instead of two-tailed ones. Now the intervals computation
should be fixed for 90%, 95% and 99%.