First beta of LaTe with additional AMQP 1.0 support and code re-organization
This commit introduces an optional support for the AMQP 1.0 protocol, using
the Apache Qpid Proton library, other than the standard LaMP over UDP tests.
At the moment AMQP 1.0 is supported only for unidirectional tests (i.e. from
producer, to broker, to consumer) and User-to-user latency, as Qpid Proton
does not allow, as far as we understood, direct control over the underlying
socket descriptor used to send packets. This will be better investigated in
the near future.

This commit also introduces some code re-organization, to improve re-usability
of some parts of the code and to make the development of new features a bit
easier in the future.

Two new options (other than -a to test over AMQP 1.0) have been added, namely:
-V: toggle a work-in-progress verbose mode (additional [INFO] messages are printed)
-V is enabling additional messages only when testing over AMQP 1.0, at the moment
-T: manually set up client timeout, instead of using the default value

When testing over AMQP 1.0, also the -H option, to specify the broker address when
in server/consumer mode (-s), has been added.

The Makefile is still work in progress and may be a bit broken. A fix will be committed
very soon, which should allow to compile LaTe with UDP support only (for instance when
no Qpid Proton library is available on the target system) or with full AMQP 1.0+UDP
support.