Fixing Makefile for the new AMQP 1.0 option
Now the Makefile can be used to compile both the standard version
of LaTe (LaMP over UDP support, for the time being) and the one
supporting both LaMP over UDP and LaMP over AMQP 1.0.
The default target is 'compilePC', for compiling the UDP only version.
The version with full AMQP 1.0+UDP support can be compiled with
'compilePCfull' or 'compilePCfulldebug'.
A temporary 'LaTe-full' binary file will be generated, which will be
then renamed to 'LaTe' when the compilation is completed.
The choice of letting the user choose between UDP or UDP+AMQP 1.0 during
compilation is due to the fact that some use cases may require only UDP
and an executable which is as much as lightweight as possible (thus without
all the code related to other protocols relying on external libraries, such
as AMQP which requires Qpid Proton to be installed).