Fix [reception timeout] print and add -N option
This commit fixed the [reception timeout] information which is printed
by a client, to take into account the new -T client-only option, and
introduces the -N options, to force LaTe to bind to all the local
interfaces instead of looking for a specific wired or wireless
interface. -N should be used as an alternative to -e.