Implementing -W on server and improving -W file creation
This commit implements the -W option also for servers, when working
in unidirectional mode.
It also improves the -W file creation, which will now check if the
specified .csv file already exists. If yes, it will try to create a
new file with name <original_name>_0001.csv. If also this file already
exists, it will try to create <original_name>_0002.csv, and so on, until
9999 attemps are performed. When all the available attemps are performed,
it will simply append to <original_name>.csv.