Displaying negative latency values during unidirectional tests
Now LaTe, in case a wrong negative unidirectional latency is computed,
due to bad clock synchronization, will display the actual negative
value instead of just considering it as '0 ms'.
This also affects the data saved to "per-packet data" CSV files (-W):
if the latency value is negative, it will be saved with the actual
negative value instead of '0', setting, in any case, the 'Error' column
to 1.