Save random periodicity type and parameter to -f CSV file
This commit updates report_manager.c to print the random periodicity
distribution type and parameter to a CSV file, for each test, when
selecting one with -f and when -R is used.
When a normal fixed periodicity is used, 'fixed periodic' will be
printed inside the line of the CSV file corresponding to the current
test.
When not using -R, the batch size which is printed when any client
is launched will be now equal to the total number of packets, instead
of being printed as '1', which seem to make more sense, as, when
not using -R, there is no periodic interval 'batch size' at all.