Introducing the -S option to specify interface name instead of using -e/-I or -N
This commit introduces a new inyerface selection option, other than the default
behaviour (look for 802.11 interfaces), -e/-I (look for wired and other interfaces),
-N (bind to any interface).
This option (-S) can be used to provide LaTe the name of an interface to be used,
instead of automatically looking for available interfaces and setting one based
on a index which is displayed when using -h.
It can be useful to select AF_INET, non-tun, interfaces which are not listed when
using -h (such as in the case of certain data interfaces on Android mobile phones).