First implementation of timeout management and cyclic number reconstruction
This commit introduces the first implementation of the following features:
1) Now a timeout won't prevent LaTe to generate and print/save the report
In case a timeout occurs, the report will also contain the last sequence
number before the timeout occurred and the packet loss up to that sequence
number.
The report structure has been updated accordingly, adding few more useful
fields, such as the last sequence number and a flag indicating whether
a timeout occurred or not.
2) LaTe will now attempt to detect each time sequence numbers are reset,
due to being cyclical 16-bit integers, i.e. each time LaMP sequence numbers
are switching 65535->0, and to reconstruct the "full" sequence numbers
by counting how many times a reset occur during a test.
These reconstructed sequence numbers are managed in report_manager and are
printed to a CSV file when -W is speficied.

This implementation is still undergoing some testing.