Improving the options managements and adding the 'w' option
This commit improves the options management, by defining macros
for each valid option, which can then be used to more easily
add their description inside print_long_info().
This commit also introduces, inside the options module, a new
'w' option, which can be used to send the per-packet data to
a UDP socket.
This option, at the moment, can be read, but the UDP socket
functionality is still not implemented. It will be implemented
in a future commit which is expected to be pushed later today or
tomorrow.