Implementation of -w UDP socket to send per-packet data to external applications
This commit implements the first version of the -w option, which
can be used, like -W for the CSV files, to send the per-packet
measurement data over a UDP socket to any listening application,
which can receive and then further process the data coming from
LaTe.
This feature is expected to be modified and further enhanced in
the next commits, so it should be still considered as experimental.