Final implementation of the -w socket for sending data to external applications
This commit introduces the final implementation of the -w option.
Now the packets containing the test data, that can be received
by an external application for further processing, involve both
the per-packet data and the final test data (avg, min, max, ...).
Two parallel sockets are now involved: a UDP socket for sending
per-packet data and a TCP socket for more sensitive data, like
the final test data, which is delivered inside a single packet.
They both refer to the same IP address and port, as specified
with -w.
The description of the -w option when using '-h' should better
describe how this is supposed to work.
This commit also fixes few small bugs and typos throughout the code,
while improving the error management (now the return values of the
read() calls for the timer are taken into account, and warnings should
be printed if something is going wrong).