Implementing native support for sending metrics to Graphite
This commit introduces a new option (-g), which can be used
to aggregate and periodically send latency-related metrics
to an installation of Graphite, by relying on the Carbon
plaintext protocol.
The options.c file has been updated in order to provide
a description on how to use the new option, when LaTe is
called with -h.
The code for parsing the "<IPv4:port>,interface" string,
when -g or -w are used, has been moved to a separate function
too.

This commit also fixes a bug in which, when using raw sockets (-r)
together with the -S option, the source MAC address wasn't
properly set in all the outgoing packets.