Introducing the -i option (test duration)
Introducing the -i option, by means if which it should be possible
to specify the test duration in seconds, instead of relying on the
total number or sent packets, specified with -n.
If both -n and -i are specified and -t is not specified, LaTe will
automatically try to compute the best periodic interval between
packets to respect both the requested -n and -i (giving priority
to -n).
-t, -n and -i cannot be specified together.
This commit also updates the -f and -w reporting functions to send/
save this new information about the test duration, if a client was
started with the '-i' option.