New option (-z) to terminate the test at a given hour:minute:second
This commit introduces a new option (-z), working in a similar
way as the recently introduced "-i", but used to make the current
test last until the specified hour:minute:second.
When it is specified, a test duration is computed, just before
starting the test duration timer, by making the difference between
the hh:mm:ss specified with -z and the current local time.
Then, this computed value is saved inside options->duration_interval,
as if it was specified with -i.
The function performing the test duration computation has been
added to options.c/.h (setTestDurationEndTime()) and it is now
being called by the UDP clients/AMQP producer when -z is specified.