Introduce options to force UDP source/destination ports
This commit introduces two new (long) options:
--udp-force-src-port <port>: to force, in the client, the usage of
a specific UDP source port (instead of letting the OS choose one)
--udp-force-dst-port <port>: to force, in the server, the usage of
a specific UDP destination port (instead of using the source port
of the received packets, as in standard UDP communications)
These options can be used, for the time being, only with non-raw
sockets (i.e. when -r is not specified).