Add --bind-to-ip option
This commit adds a new option (--bind-to-ip) to bind to a specific
IP address, instead of specifying an interface name or an interface
internal index.
This option works like -B in iperf 2, and it can be useful when
testing over an interface on which IP Aliases are in use (i.e.,
an interface with multiple IP addresses).
Indeed, -S will result in binding to the first IP address of the
given interface, if more than one is available, which is not
always the desired behaviour when aliases are used.